+++ Spec/Programs/JobQueue.xs
@@ -6,7 +6,7 @@
 
 
 
-import optional, process, sync, stdio;
+import process, sync, stdio;
 
 data Job {
     id: Int;